fixes for common changes
